forecastapp_novenv/requirements.txt
Flask==2.2.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0